Port xf86-video-v4l driver to V4L2
V4L1 API was dropped on kernel 2.6.38. Even the emulation layer
inside kernel was dropped. While it might still be possible to use
X with libv4l and a LD_PRELOADER setup, the proper way is to port
the driver to use the V4L2 API.

This change brings some improvements to the code, as now all
integer/boolean/menu controls are visible, and all video standards.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>