Provide a more consistent message if FBUF fails
Instead of providing an useless info that FBUF fails, print an
error indicating what's the Fourcc format that fails.

Also, on debug mode, prints a message indicating the selection of a
given fourcc. This helps to verify if both the video adapter and the
video capture board were set to the same video format.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>