Fix arguments for v4l_check_yuv
The device name and the file descriptors are wrong at the cal for
v4l_check_yuv(). Due to that, the driver were falling back to work
with RGB formats.

While here, add a code to prevent V4L driver to work if no video
adapter provide Xv overlay method.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>